Moved cell headers into a separate file
